fix: ensure tiled_dit output is float32 for ComfyUI compatibility
- Convert tiled output to float32 after final division
- Matches non-tiled path which uses tensor2video() with .float()
- Fixes incompatibility with downstream nodes like Fill's FILM
- Ensures all IMAGE tensors are float32 as required by ComfyUI
- Resolves dtype mismatch only present in tiled_dit mode

This fixes the issue where tiled processing outputs bf16/fp16 tensors
that cause failures in nodes expecting float32 IMAGE tensors.

+++ nodes.py
@@ -304,7 +304,7 @@ def flashvsr(pipe, frames, scale, color_fix, tiled_vae, tiled_dit, tile_size, ti
             clean_vram()
             
         weight_sum_canvas[weight_sum_canvas == 0] = 1.0
-        final_output = final_output_canvas / weight_sum_canvas
+        final_output = (final_output_canvas / weight_sum_canvas).float()
     else:
         log("[FlashVSR] Preparing frames...")
         LQ, th, tw, F = prepare_input_tensor(_frames, _device, scale=scale, dtype=dtype)
